requirement.txt
streamlit == 1.40.2
ollama == 0.4.8
PyPDF == 3.0.1
